backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
shap
stable-baselines3
gymnasium
requests
reportlab
httpx
python-dotenv
